+++ src/main/java/jcifs/dcerpc/msrpc/eventing/LogonEvent.java
@@ -59,7 +59,7 @@ public class LogonEvent extends Event {
     public final String elevatedToken;
 
     protected LogonEvent(int eventId, int version, int level, int task, int opcode, String keywords,
-               String timeCreated, int eventRecordId, String activityId, int processId, int threadId,
+               String timeCreated, long eventRecordId, String activityId, int processId, int threadId,
                String channel, String computer, Map<String, String> eventData) {
         super(eventId, version, level, task, opcode, keywords, timeCreated, eventRecordId, activityId,
                 processId, threadId, channel, computer);
